requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
jinja2
python-telegram-bot==13.15
